Rename in order to simplify HelperClassPredicate
