+++ instrumentation-appender-api-internal/src/main/java/io/opentelemetry/instrumentation/api/appender/internal/NoopLogBuilder.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.appender.api.internal;
+package io.opentelemetry.instrumentation.api.appender.internal;
 
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.context.Context;
