+++ instrumentation-appender-api-internal/src/test/java/io/opentelemetry/instrumentation/api/appender/internal/AgentLogEmitterProviderTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.appender.api.internal;
+package io.opentelemetry.instrumentation.api.appender.internal;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
