+++ instrumentation-appender-sdk-internal/src/main/java/io/opentelemetry/instrumentation/sdk/appender/internal/DelegatingLogBuilder.java
@@ -3,12 +3,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.appender.sdk.internal;
+package io.opentelemetry.instrumentation.sdk.appender.internal;
 
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.context.Context;
-import io.opentelemetry.instrumentation.appender.api.internal.LogBuilder;
-import io.opentelemetry.instrumentation.appender.api.internal.Severity;
+import io.opentelemetry.instrumentation.api.appender.internal.LogBuilder;
+import io.opentelemetry.instrumentation.api.appender.internal.Severity;
 import java.time.Instant;
 import java.util.concurrent.TimeUnit;
