+++ instrumentation/log4j/log4j-appender-1.2/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/log4j/appender/v1_2/Log4jHelper.java
@@ -8,9 +8,9 @@ package io.opentelemetry.javaagent.instrumentation.log4j.appender.v1_2;
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.api.common.AttributesBuilder;
 import io.opentelemetry.context.Context;
-import io.opentelemetry.instrumentation.appender.api.internal.AgentLogEmitterProvider;
-import io.opentelemetry.instrumentation.appender.api.internal.LogBuilder;
-import io.opentelemetry.instrumentation.appender.api.internal.Severity;
+import io.opentelemetry.instrumentation.api.appender.internal.AgentLogEmitterProvider;
+import io.opentelemetry.instrumentation.api.appender.internal.LogBuilder;
+import io.opentelemetry.instrumentation.api.appender.internal.Severity;
 import io.opentelemetry.semconv.trace.attributes.SemanticAttributes;
 import java.io.PrintWriter;
 import java.io.StringWriter;
