+++ instrumentation/log4j/log4j-appender-2.16/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/log4j/appender/v2_16/Log4jHelper.java
@@ -5,8 +5,8 @@
 
 package io.opentelemetry.javaagent.instrumentation.log4j.appender.v2_16;
 
-import io.opentelemetry.instrumentation.appender.api.internal.AgentLogEmitterProvider;
-import io.opentelemetry.instrumentation.appender.api.internal.LogBuilder;
+import io.opentelemetry.instrumentation.api.appender.internal.AgentLogEmitterProvider;
+import io.opentelemetry.instrumentation.api.appender.internal.LogBuilder;
 import io.opentelemetry.instrumentation.log4j.appender.v2_16.internal.ContextDataAccessor;
 import io.opentelemetry.instrumentation.log4j.appender.v2_16.internal.LogEventMapper;
 import java.util.Map;
