+++ instrumentation/logback/logback-appender-1.0/library/src/main/java/io/opentelemetry/instrumentation/logback/appender/v1_0/OpenTelemetryAppender.java
@@ -7,9 +7,9 @@ package io.opentelemetry.instrumentation.logback.appender.v1_0;
 
 import ch.qos.logback.classic.spi.ILoggingEvent;
 import ch.qos.logback.core.AppenderBase;
-import io.opentelemetry.instrumentation.appender.api.internal.LogEmitterProvider;
-import io.opentelemetry.instrumentation.appender.api.internal.LogEmitterProviderHolder;
-import io.opentelemetry.instrumentation.appender.sdk.internal.DelegatingLogEmitterProvider;
+import io.opentelemetry.instrumentation.api.appender.internal.LogEmitterProvider;
+import io.opentelemetry.instrumentation.api.appender.internal.LogEmitterProviderHolder;
+import io.opentelemetry.instrumentation.sdk.appender.internal.DelegatingLogEmitterProvider;
 import io.opentelemetry.instrumentation.logback.appender.v1_0.internal.LoggingEventMapper;
 import io.opentelemetry.sdk.logs.SdkLogEmitterProvider;
